added file to solve package installation issues, added possible fix for qgstask not running on first call
